refactor: move test utilities and mocks to tests directory
- Move all test utilities from src/ to tests/test-utils/
- Move all mock services from packages/core/mocks/ to tests/mocks/
- Update all import paths to reference new locations
- Fix test isolation issues in migration tests by using unique temp directories
- Update bunfig.toml with proper coverage exclusion patterns
- Add QA review documentation to story file

This change improves code organization by clearly separating production
code from test code, making the codebase structure more maintainable.
