+++ docs/stories/1.13.ioc-dependency-injection.story.md
@@ -323,6 +323,7 @@ afterEach(() => {
 | 2025-01-09 | 1.0 | Initial story creation for IoC/DI implementation | Bob (SM) |
 | 2025-01-09 | 1.1 | Enhanced with phased migration, rollback strategy, debugging capabilities, and lifecycle hooks | Sarah (PO) |
 | 2025-01-09 | 1.2 | Fixed BaseService reference (line 7, not 161); Added Testing subsection per template requirements | Sarah (PO) |
+| 2025-09-09 | 1.3 | QA Review completed - Addressed minor gaps: confirmed ESLint working, verified test coverage, initiated StrykerJS mutation testing | James (Dev Agent) |
 
 ## Dev Agent Record
 
@@ -332,6 +333,11 @@ claude-opus-4-1-20250805
 ### Debug Log References
 - Task execution tracked via TodoWrite tool
 - No errors encountered during implementation
+- QA Review completed via *review-qa command
+- ESLint configuration verified and working
+- Test coverage measurement completed successfully
+- StrykerJS mutation testing initiated (running in background)
+- All quality checks pass (lint, format, typecheck)
 
 ### Completion Notes List
 - All Tasks 1-10 completed successfully
@@ -355,6 +361,14 @@ Performance Highlights:
 - Over 1M operations/second for singleton resolution
 - Memory efficient with ~0.07KB per operation
 - Better performance than industry-standard DI frameworks
+
+QA Review Results:
+- ESLint configuration confirmed working (existing)
+- All quality checks pass: lint, format, typecheck
+- Test coverage measurement completed successfully
+- All tests passing (42+ tests)
+- StrykerJS mutation testing initiated and running
+- QA gaps addressed: ESLint working, tests verified
 
 ### File List
