+++ packages/core/tests/integration/DIMigration.test.ts
@@ -2,11 +2,11 @@ import { describe, test, expect, beforeEach, afterEach } from 'bun:test';
 import { ServiceProvider } from '../../src/container/ServiceProvider';
 import { CompatibilityLayer } from '../../src/container/CompatibilityLayer';
 import { FeatureFlagManager } from '../../src/container/FeatureFlags';
-import { MockLoggerService } from '../../mocks/LoggerService.mock';
-import { MockConfigService } from '../../mocks/ConfigService.mock';
-import { MockFileSystemService } from '../../mocks/FileSystemService.mock';
-import { MockStateManagerService } from '../../mocks/StateManagerService.mock';
-import { MockWorkflowEngineService } from '../../mocks/WorkflowEngineService.mock';
+import { MockLoggerService } from '../mocks/LoggerService.mock';
+import { MockConfigService } from '../mocks/ConfigService.mock';
+import { MockFileSystemService } from '../mocks/FileSystemService.mock';
+import { MockStateManagerService } from '../mocks/StateManagerService.mock';
+import { MockWorkflowEngineService } from '../mocks/WorkflowEngineService.mock';
 
 describe('DI Migration Integration', () => {
   let provider: ServiceProvider;
