+++ packages/core/tests/mocks/index.ts
@@ -1,5 +1,5 @@
-export { MockLoggerService, type LogEntry } from '../../mocks/LoggerService.mock';
-export { MockStateManagerService } from '../../mocks/StateManagerService.mock';
-export { MockWorkflowEngineService } from '../../mocks/WorkflowEngineService.mock';
-export { MockConfigService } from '../../mocks/ConfigService.mock';
-export { MockFileSystemService } from '../../mocks/FileSystemService.mock';
+export { MockLoggerService, type LogEntry } from './LoggerService.mock';
+export { MockStateManagerService } from './StateManagerService.mock';
+export { MockWorkflowEngineService } from './WorkflowEngineService.mock';
+export { MockConfigService } from './ConfigService.mock';
+export { MockFileSystemService } from './FileSystemService.mock';
