+++ packages/core/tests/test-utils/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Test utilities for @checklist/core
+ * These utilities are for testing purposes only
+ */
+
+export { MockLogger, InMemoryLogger } from './MockLogger';
+export { TestDataFactory } from './TestDataFactory';
+export { LogAssertions } from './LogAssertions';
